backend/requirements.txt

django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
psycopg2
requests

Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.0

psycopg2-binary==2.9.10

sqlparse==0.5.3
typing_extensions==4.14.0


langchain
pillow
transformers
salesforce-lavis
requests
groq
langchain-community